% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsyllable.R
\docType{data}
\name{data_int_syllables}
\alias{data_int_syllables}
\title{\code{data_int_syllables} provides an English-language syllables dictionary;
it is an integer vector whose element names correspond to English words.}
\format{
An object of class \code{integer} of length 133245.
}
\usage{
data_int_syllables
}
\description{
\code{data_int_syllables} provides an English-language syllables dictionary;
it is an integer vector whose element names correspond to English words.
}
\keyword{datasets}
