% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_syllables_en}
\alias{data_syllables_en}
\title{Syllable counts of English words}
\format{
An object of class \code{integer} of length 125698.
}
\source{
Version 0.7b of the CMU Pronouncing Dictionary.  See
\url{https://github.com/cmusphinx/cmudict}.
}
\usage{
data_syllables_en
}
\description{
A named integer vector of syllable counts for English words. Based on a
pronunciation dictionary for North American English that contains over
134,000 words and their pronunciations, from the \href{https://github.com/cmusphinx/cmudict}{Carnegie Mellon University Pronouncing Dictionary}.
}
\note{
\code{data_syllables_en} is a data object consisting of a named numeric vector
of syllable counts for the words used as names.  This is the default object
used to count English syllables.  For words with multiple pronunciation
variants, we use the first entry.

This object that can be accessed directly, but we strongly encourage you to
access it only through the \code{\link[=nsyllable]{nsyllable()}} wrapper function.
}
\keyword{datasets}
