% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsyllable-package.R
\docType{package}
\name{nsyllable-package}
\alias{nsyllable-package}
\title{nsyllable: Count syllables in character vectors}
\description{
Counts syllables from character vector inputs.  Imputes syllables as the
number of vowel sequences for words not found.
}
\seealso{
Useful links:
\itemize{
\item \url{https://github.com/quanteda/nsyllable}
\item Report bugs at \url{https://github.com/quanteda/nsyllable/issues}
}
}
\author{
Kenneth Benoit
}
